+++ lib/ui/component/snackBar/error_snackbar.dart
